# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The REST API a Quark device serves to its Flutter clients. Every endpoint except\n/auth/setup, /auth/login, /auth/recover, /auth/request-account and /auth/status needs a\nsession token. Sign in with POST /auth/login, then click Authorize and enter the word\nBearer, a space, and the token.",
        "title": "Quark API",
        "contact": {},
        "version": "v0"
    },
    "basePath": "/api/v0",
    "paths": {
...
        "/albums": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the caller's own photo albums as a flat list, creating their Favorites album on first use. Admins see only their own albums too. Use ?tree=true to get a nested tree.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "List all photo albums",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Return as nested tree (default false)",
                        "name": "tree",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/v0_albums.AlbumJSON"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a photo album owned by the caller, optionally nested under one of the caller's albums. The name cannot contain / and must be unique among its siblings ignoring case; the caller's root albums, their Favorites album included, are siblings of each other.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Create a photo album",
                "parameters": [
                    {
                        "description": "Album name and optional parent ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v0_albums.createAlbumRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/v0_albums.AlbumJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request: missing name, a / in the name, or parent not found among the caller's albums",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden: system album as the parent",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict: an album with that name already exists here",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
...
        "/albums/{id}/move": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Changes the parent of one of the caller's albums. Pass null parentId to move to root. The new parent must be the caller's own and must not already hold an album with the same name ignoring case; the caller's root albums, their Favorites album included, are siblings of each other.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Move a photo album to a new parent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Album ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New parent ID (null for root)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v0_albums.moveAlbumRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v0_albums.AlbumJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request: invalid id or body, or parent not found among the caller's albums",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden: system album, or a system album as the parent",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found: no album of the caller's has that id",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict: an album with that name already exists here",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
...
        "/auth/sessions/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a specific session. The session ID is the SHA-256 hex hash of the session token, as returned by GET /auth/sessions.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Revoke a session by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID (hex SHA-256 of token)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "revoked": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
...
        "/photos/favorite": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns whether the specified photo is in the caller's own favorites. Needs read access on the photo.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "favorites"
                ],
                "summary": "Check if a photo is favorited",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device serial",
                        "name": "serial",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Relative path to the photo",
                        "name": "relPath",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v0_favorites.favoriteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adds the photo to the caller's own favorites if not already favorited; removes it otherwise. Needs read access on the photo.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "favorites"
                ],
                "summary": "Toggle a photo favorite",
                "parameters": [
                    {
                        "description": "Photo reference",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v0_favorites.favoriteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v0_favorites.favoriteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/serverutil.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "serverutil.ContentType": {
            "type": "string",
            "enum": [
                "text/html",
                "application/json"
            ],
            "x-enum-varnames": [
                "ContentTypeHTML",
                "ContentTypeJSON"
            ]
        },
        "serverutil.Response": {
            "type": "object",
            "properties": {
                "contentType": {
                    "$ref": "#/definitions/serverutil.ContentType"
                },
                "data": {},
                "error": {},
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "v0_albums.AlbumJSON": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v0_albums.AlbumJSON"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "itemCount": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                },
                "smartType": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "v0_albums.createAlbumRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                }
            }
        },
        "v0_albums.moveAlbumRequest": {
            "type": "object",
            "properties": {
                "parentId": {
                    "type": "integer"
                }
            }
        },
...
        "v0_favorites.favoriteRequest": {
            "type": "object",
            "properties": {
                "deviceSerial": {
                    "type": "string"
                },
                "relPath": {
                    "type": "string"
                }
            }
        },
        "v0_favorites.favoriteResponse": {
            "type": "object",
            "properties": {
                "isFavorite": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "A session token from POST /auth/login, entered as the word Bearer, a space, and the token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}